ai_openai_whisper/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==4.4.2
exceptiongroup==1.1.3
fastapi==0.103.2
ffmpeg-python==0.2.0
filelock==3.12.4
future==0.18.3
h11==0.14.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
packaging==23.2
panda==0.3.1
pandas==2.1.1
Pillow==10.0.1
pluggy==1.3.0
proglog==0.1.10
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytube==15.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.3.3
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
